+++ octet/src/containers/string.h
@@ -187,7 +187,7 @@ namespace octet { namespace containers {
 
     void vformat(const char *fmt, va_list v) {
       #ifdef WIN32
-        int cur_len = strlen(data_);
+        size_t cur_len = strlen(data_);
         int len = _vscprintf(fmt, v);
         if (len) {
           if (cur_len) {
